Fixes polygon bounding box calculations and adds traingle and square
Test cases need to be written for triangle and square and their PS implementation needs to take into account variables rather than hard coded center points. Default ctors need to be made for triangle and square too. The circle printing test fails for some reason possibly due to the struct changes in shape, the error is contained within circle's PostScriptRepresentation function

+++ include/Shape.h
@@ -15,13 +15,13 @@
 namespace PostLib {
     
     typedef struct{
-        signed int x;
-        signed int y;
+        double x;
+        double y;
     }PostScriptPoint;
     
     typedef struct{
-        signed int width;
-        signed int height;
+        double width;
+        double height;
     }Size;
     
     typedef struct{
